Replace "de:" with "rl:" for RelicLink ID changes
Changes are being made to be able to support more games of the franchise using the RelicLink API (e.g. AoE2DE, AoE3DE, AoE4).

Players on this platform have a game independent ID. This change lets us add more players of the different games of the franchise in the future.